Refactor phase 3b: split routes_public into a per-surface package
webapp/routes_public.py was 2,849 lines with exactly one top-level
definition: its whole body was a single 2,779-line register(app, logger)
with all 21 route handlers nested inside it. No module-level split was
possible, and no surface — landing pages, sitemap, stats, alerts, logs —
could be read or tested apart from the others.

That structure made the split simpler rather than harder. Every handler
closes over only `app` (for the @app.route decorator) and `route_logger`,
verified by walking the AST for names resolving to register's scope. So
each surface keeps its own register(app, route_logger) and the handler
bodies move verbatim — still nested inside a register, just a much
smaller one. No reindentation, no rebinding, no signature changes.

  pages.py       189  landing, about/help, policy docs, health, monitor
  sitemap.py     115  /sitemap.xml
  stats.py       693  /stats
  alerts.py      648  /alerts and its PDF export
  logs_data.py  1116  _load_logs_data, via build_logs_loader(route_logger)
  logs.py        265  /logs and its CSV/PDF exports

_load_logs_data is a helper the three /logs handlers share, not a route.
A factory keeps its 1,057-line body byte-identical — only the enclosing
scope changed — instead of re-signaturing it to take route_logger. logs.py
receives it as a parameter named _load_logs_data so the call sites resolve
unchanged.

webapp/routes_public.py stays as a 31-line shim re-exporting register, so
the route-module registry in webapp/__init__.py is untouched. Its __all__
is what keeps ruff --fix from stripping the re-export; unlike an
__init__.py it gets no F401 exemption.

Verification, each confirmed discriminating before being trusted:
- 21/21 handlers ast.dump()-identical; every non-blank line of the
  original register() placed in exactly one module (2,753 lines, 0
  unplaced).
- Flask URL map diffed against a worktree at the pre-split commit: 549
  rules, 0 differences. Mutation-checked by deleting one register() call,
  which the diff caught.
- All 28 public surfaces fetched through the test client on both sides
  and hashed with volatile content scrubbed: 28/28 identical, 28 distinct
  digests.

logs_data.py, stats.py and alerts.py remain over the 400-line guidance —
each is dominated by one enormous function, which module-level splitting
cannot shrink. Opened as phase 3b-ii in the plan.

Full suite: 2045 passed, 60 xfailed, 7 xpassed.
